refactor: 상품 관리 페이지 컴포넌트 역할 재분배
